rezerwacje mostly fine

+++ RejestracjaSal/Models/Reservations.cs
@@ -7,7 +7,7 @@ namespace RejestracjaSal.Models
         [Key]
         public int Reservation_id { get; set; }
         public int User_id {  get; set; }
-        public int Total_price { get; set; }
+        public float Total_price { get; set; }
         public bool Paid {  get; set; }
         public string? Comments {  get; set; }
     }
